requirements.txt
# Core dependencies
python-dotenv==1.0.1

# HTTP client for API calls
requests==2.31.0

# OpenAI SDK (works with NVIDIA NIM and OpenRouter)
openai==1.12.0

# HuggingFace Inference API
huggingface-hub==0.20.3

# Rich for CLI colors
rich==13.7.0